Separated validator class to reduce confusion in controller-support
