+++ controllers/controller-support.js
@@ -1,6 +1,5 @@
 // CRUD GENERATOR!!!
-// Controller to perform CRUD
-const validateDate = require("validate-date");
+// Controller to perform CRUD and other support functions 
 const { isValidObjectId } = require("mongoose");
 const User = require('../models/user');
 const Contact = require('../models/contact');
@@ -87,90 +86,6 @@ function findOne(controller, req, res) {
     });
 }
 
-// Checks for valid character in fields such as names ==============================
-function checkInvalid(string) {
-  var format = /[!@#$%^&*()_+\-=\\[\]{};':"\\|,.<>\\/?]+/;
-
-  if (format.test(string)) {
-    return true;
-  } else {
-    return false;
-  }
-}
-
-// Checks for valid dateTime =======================================================
-function checkValidDate(date) {
-  return validateDate(date);
-}
-
-// Function to check for valid Ids 
-// Used mostly for user and contact
-async function checkValidId(controller, id) {
-  var check = true;
-  if (!id || !isValidObjectId(id)) { check = false; }
-
-  await controller.findById(id).then((foundId) => {
-    if (!foundId) {
-      check = false;
-    }
-  });
-  return check;
-}
-
-// FUnction to check for existence of data block =====================================
-// Used mostly for convo and relationship
-async function checkExist(controller, ids) {
-  var check = false;
-  await controller.findOne({ people: ids }).then((found) => {
-    if (found) {
-      check = true;
-    }
-  });
-  return check;
-}
-
-// FUnction to check for valid contactIds with given belongsTo
-// Used mostly for user and contact
-async function checkValidIdWithBelongsTo(controller, id, belongsTo) {
-  var check = true;
-  if (!id || !isValidObjectId(id)) { check = false; }
-
-  await controller.findById(id).then((foundId) => {
-    if (!foundId || foundId.belongsTo != belongsTo) {
-      check = false;
-    }
-  });
-  return check;
-}
-
-// Check for self-existence in the database ==========================================
-// Basic checking: used for relationship and convo
-async function validRelationshipOrConvo(controller2, req, type) {
-  const sortedIds = req.body.people.sort();
-  // Check for duplicate userIds
-  if (sortedIds[0] == sortedIds[1]) {
-    return false;
-  }
-
-  var firstValid, secValid;
-
-  // Then check for valid Ids
-  if (type == 'relationship') {
-    firstValid = await checkValidIdWithBelongsTo(Contact, sortedIds[0], req.body.belongsTo);
-    secValid = await checkValidIdWithBelongsTo(Contact, sortedIds[1], req.body.belongsTo);
-  }
-  else {
-    firstValid = await checkValidId(User, sortedIds[0]);
-    secValid = await checkValidId(User, sortedIds[1]);
-  }
-  // Check for existence
-  const existed = await checkExist(controller2, sortedIds);
-  if (firstValid && secValid && !existed) {
-    return true;
-  }
-  return false;
-}
-
 // Get all contact/relationships that belong to a specific user ======================
 async function getAllByUserId(controller, req, res) {
   const ownerId = req.params.belongsToId;
@@ -218,7 +133,6 @@ async function getNameFromContactId(belongsTo, list) {
 }
 
 module.exports = {
-  updateData, deleteData, findAllData, findOne,
-  checkInvalid, checkValidDate, validRelationshipOrConvo,
-  checkValidId, getAllByUserId, getNameFromContactId,
+  updateData, deleteData, findAllData,
+  findOne, getAllByUserId, getNameFromContactId,
 };
